chore: rename StatusAssertionScreen to CredentialStatusScreen
